+++ graphql/fetchUserProfileCard.ts
@@ -1,5 +1,16 @@
 import { gql } from "@apollo/client";
+import React from "react";
+export type Users = {
+  id: string;
+  name: string;
+  profile_image: string;
+  badges_earned: number;
+  created_at: string;
+};
 
+export type FetchUserProfileCardResponse = {
+  users: Array<Users>;
+};
 export const FETCH_USER_PROFILE_CARD = gql`
   query fetchUsers {
     users(where: { enrolled: { _eq: true } }) {
